==> Scripts/02_preprocess.ipynb <==
import os
import wandb
import pandas as pd
import numpy as np
import json
import warnings
import util as ut
warnings.filterwarnings('ignore')

run = wandb.init(project=ut.PROJECT_NAME, entity=ut.ENTITY, job_type="Upload",
                 notes="Uploading the Dream11 dataset (full) with 5 classes")

path = '../Inputs/ball-by-ball prediction/'
# wandb.log_artifact(path+'clean_data.csv', name='clean_data', type='ball_by_ball_prediction')
# run.finish()

clean_data = pd.read_csv('../Inputs/all_matches_cleaned.csv', index_col=0)

def get_stats(df, till_date, typ="bowler"):
    # Get the data till the given date
    df["start_date"] = pd.to_datetime(df["start_date"])
    df = df[df["start_date"] < till_date]

    # Get the number of runs scored by each batsman
    runs = df.groupby(typ)["runs_off_bat"].value_counts().unstack(level=1)
    runs = runs.dropna(axis=1, thresh=runs.shape[0] * 0.1).fillna(0)
    runs.columns = ["0_runs", "1_runs", "2_runs", "3_runs", "4_runs", "6_runs"]

    # Get the number of balls faced by each batsman
    balls = pd.DataFrame(df.groupby(typ)["ball"].count())

    # Get the number of wides faced by each batsman
    df.loc[:, "wides"] = (
        df["wides"].fillna(0).apply(lambda x: 0 if x == 0 else 1).tolist()
    )
    wides = pd.DataFrame(df.groupby(typ)["wides"].sum())

    # Get the number of dismissals faced by each batsman
    dismissals = pd.DataFrame(
        df.groupby(typ)["wicket_type"].apply(lambda x: x.notna().sum())
    ).rename(columns={"wicket_type": "num_dismissals"})

    # Add the number of balls faced and the number of wides faced to get the total number of balls faced
    balls["total_balls"] = balls["ball"] + wides["wides"]
    runs["0_runs"] = runs["0_runs"].subtract(dismissals["num_dismissals"])

    # Merge all the dataframes
    stats = pd.concat([runs, dismissals, wides, balls["total_balls"]], axis=1).fillna(0)
    return stats.reset_index()

def divide_data(cleandf):
    # Divide the data into history, train, and test
    temp = cleandf.sort_values(by=['start_date', 'innings', 'overs', 'balls']).copy()
    train = temp.query('"2016-01-01" <= start_date < "2022-01-01"')
    test = temp.query('start_date >= "2022-01-01"')
    return train, test


def merge_feats(train):
    temp = train.copy()
    temp = temp.sort_values(by=['start_date', 'innings', 'overs', 'balls'])
    
    unique_dates = temp['start_date'].unique().tolist()

    train_dfs = [
        temp[temp['start_date'] == date].merge(
            get_stats(clean_data, date, 'striker').add_prefix('bat_'),
            left_on='striker',
            right_on='bat_striker',
            how='left'
        ).merge(
            get_stats(clean_data, date, 'bowler').add_prefix('bowl_'),
            left_on='bowler',
            right_on='bowl_bowler',
            how='left'
        )
        for date in unique_dates
    ]
    
    newtrain = pd.concat(train_dfs)
    newtrain = newtrain.sort_values(by=['start_date', 'innings', 'overs', 'balls'])
    newtrain = newtrain.drop(['start_date', 'ball', 'extras','noballs','byes','legbyes', \
                            'player_dismissed', 'bat_striker', 'bowl_bowler'], axis=1)
    
    return newtrain

def create_labels_1(row):
        if row['wicket_type'] != 'none':
            return "Wicket"
        elif row['runs_off_bat'] == 0:
            return "0_runs"
        elif row['runs_off_bat'] == 1:
            return "1_runs"
        elif row['runs_off_bat'] == 2:
            return "2_runs"
        elif row['runs_off_bat'] == 3:
            return "3_runs"
        elif row['runs_off_bat'] == 4 or row['runs_off_bat'] == 5:
            return "4_runs"
        elif row['runs_off_bat'] == 6 or row['runs_off_bat'] == 7:
            return "6_runs"
        elif row['wides'] != 'none':
            return "Wide"
        else:
            raise ValueError('Error in creating y, check the data')
        
def create_labels_2(row):
    if row['wicket_type'] != 'none':
        return "Wicket"
    elif row['runs_off_bat'] == 0:
        return "dot"
    elif row['runs_off_bat'] == 1 or row['runs_off_bat'] == 2 or row['runs_off_bat'] == 3 :
        return "runs"
    elif row['runs_off_bat'] == 4 or row['runs_off_bat'] == 5:
        return "four"
    elif row['runs_off_bat'] == 6 or row['runs_off_bat'] == 7:
        return "six"
    elif row['wides'] != 'none':
        return "Wide"
    else:
        raise ValueError('Error in creating y, check the data')


def make_target(df):
    cleandf = df.copy()
    create_y = cleandf[['runs_off_bat', 'wides', 'wicket_type']].fillna('none')
    
    cleandf['target'] = create_y.apply(create_labels_2, axis=1)
    cleandf = cleandf.drop(['runs_off_bat', 'wides', 'wicket_type'], axis=1)
    return cleandf.fillna(0.0)

#slow takes 13 mins
def make_features(clean_data):
    train1, test1 = divide_data(clean_data)
    train2, test2 = merge_feats(train1), merge_feats(test1)
    del train1, test1
    train, test = make_target(train2), make_target(test2)
    del train2, test2
    return train, test

clean_data.sort_values(by=['start_date', 'innings', 'overs', 'balls'], inplace=True)

df = pd.read_csv(path+'main.csv')

dict_5c = {
    '0_runs': 'norun',
    '1_runs': 'run',
    '2_runs': 'run',
    '3_runs': 'run',
    '4_runs': 'four',
    '6_runs': 'six',
    'Wicket': 'wicket',
}
df['target'] = df['target'].apply(lambda x: dict_5c[x])

df.to_csv(path+'main_5c.csv', index=False)

wandb.log_artifact(path+'main_5c.csv', name='main_5c', type='ball_by_ball_prediction')
# wandb.log_artifact(path+'test.csv', name='test_5c_split_2022on', type='ball_by_ball_prediction')
run.finish()

def get_player_dict():
    player_dict = {}
    kinds = ["ipl", "t20s"]
    for kind in kinds:
        path_json = f"../Inputs/{kind}_json"
        json_files = [
            pos_json for pos_json in os.listdir(path_json) if pos_json.endswith(".json")
        ]

        for file in json_files:
            with open(os.path.join(path_json, file)) as json_file:
                json_data = json.load(json_file)
                for key, val in json_data["info"]["registry"]["people"].items():
                    player_dict[val] = key
    return player_dict

player_dict = get_player_dict()

player_dict

artifact = wandb.Artifact(name='player_stats', type='ball_by_ball_prediction')

batter_stats = get_stats(clean_data, "2023-03-14", "striker")
batter_stats = batter_stats.applymap(lambda x: player_dict.get(x, x))

bowler_stats = get_stats(clean_data, "2023-03-14", "bowler")
bowler_stats = batter_stats.applymap(lambda x: player_dict.get(x, x))

bat_stats = wandb.Table(dataframe=batter_stats)
bowl_stats = wandb.Table(dataframe=bowler_stats)

artifact.add(bat_stats, "bat_stats")
artifact.add(bowl_stats, "bowl_stats")

run.log_artifact(artifact)
run.finish()

clean_data['match_id'] = clean_data.index.str.split('_').str[0]

clean_data.sort_values(by=['start_date', 'innings', 'overs', 'balls'], inplace=True)
# matchid = clean_data.loc[0, 'match_id']

#if innings == 1, over = 1, ball = 1, then it is the start of the innings, add a column for the score at the start of the innings
clean_data['current_score_cumsum'] = clean_data.groupby(['match_id', 'innings'])['runs_off_bat'].cumsum()
clean_data['current_score'] = clean_data.groupby(['match_id', 'innings'])['current_score_cumsum'].shift(1).fillna(0).astype(int)
clean_data['wicket_type'] = clean_data['player_dismissed'].apply(lambda x: 1 if isinstance(x,str) else 0)
clean_data['wickets_cumsum'] = (10-clean_data.groupby(['match_id', 'innings'])['wicket_type'].cumsum())
clean_data['wickets_remaining'] = clean_data.groupby(['match_id', 'innings'])['wickets_cumsum'].shift(1).fillna(10).astype(int)
clean_data.drop('wickets_cumsum', axis=1, inplace=True)


clean_data.groupby(['match_id', 'innings', 'striker'])['runs_off_bat'].value_counts().unstack(level=3).fillna(0)[:50]

clean_data.groupby(['match_id', 'striker'])['runs_off_bat'].sum().fillna(0)[:50]
